+++ src/infra/flow-system/inmem/src/flow_configuration/flow_configuration_event_store_inmem.rs
@@ -8,7 +8,6 @@
 // by the Apache License, Version 2.0.
 
 use dill::*;
-use kamu_core::DatasetIDStream;
 use kamu_flow_system::*;
 use opendatafabric::DatasetID;
 
@@ -77,6 +76,10 @@ impl EventStore<FlowConfigurationState> for FlowConfigurationEventStoreInMem {
         query: &FlowKey,
         events: Vec<FlowConfigurationEvent>,
     ) -> Result<EventID, SaveEventsError> {
+        if events.is_empty() {
+            return Err(SaveEventsError::NothingToSave);
+        }
+
         if let FlowKey::Dataset(flow_key) = query {
             let state = self.inner.as_state();
             let mut g = state.lock().unwrap();
@@ -89,13 +92,16 @@ impl EventStore<FlowConfigurationState> for FlowConfigurationEventStoreInMem {
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
+#[async_trait::async_trait]
 impl FlowConfigurationEventStore for FlowConfigurationEventStoreInMem {
     #[tracing::instrument(level = "debug", skip_all)]
-    fn list_all_dataset_ids(&self) -> DatasetIDStream {
+    async fn list_all_dataset_ids(&self) -> FailableDatasetIDStream {
+        use futures::StreamExt;
+
+        let dataset_ids = self.inner.as_state().lock().unwrap().dataset_ids.clone();
+
         // TODO: re-consider performance impact
-        Box::pin(tokio_stream::iter(
-            self.inner.as_state().lock().unwrap().dataset_ids.clone(),
-        ))
+        Box::pin(tokio_stream::iter(dataset_ids).map(Ok))
     }
 }