+++ src/infra/flow-system/inmem/tests/tests/mod.rs
@@ -7,4 +7,5 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+mod test_flow_configuration_event_store_inmem;
 mod test_flow_event_store_inmem;