+++ src/infra/task-system/sqlite/src/task_system_event_store_sqlite.rs
@@ -76,7 +76,7 @@ impl EventStore<TaskState> for TaskSystemEventStoreSqlite {
             .map_err(|e| GetEventsError::Internal(e.int_err()));
 
             while let Some((event_id, event)) = query_stream.try_next().await? {
-                yield(Ok((event_id, event)));
+                yield Ok((event_id, event));
             }
         })
     }
@@ -86,6 +86,10 @@ impl EventStore<TaskState> for TaskSystemEventStoreSqlite {
         _task_id: &TaskID,
         events: Vec<TaskEvent>,
     ) -> Result<EventID, SaveEventsError> {
+        if events.is_empty() {
+            return Err(SaveEventsError::NothingToSave);
+        }
+
         let mut tr = self.transaction.lock().await;
         let connection_mut = tr.connection_mut().await?;
 
@@ -101,7 +105,7 @@ impl EventStore<TaskState> for TaskSystemEventStoreSqlite {
         );
 
         query_builder.push_values(events, |mut b, event| {
-            let event_task_id: i64 = (event.task_id()).into();
+            let event_task_id: i64 = event.task_id().into();
             b.push_bind(event_task_id);
             b.push_bind(event.dataset_id().map(ToString::to_string));
             b.push_bind(event.event_time());
@@ -115,15 +119,10 @@ impl EventStore<TaskState> for TaskSystemEventStoreSqlite {
             .build_query_as::<ResultRow>()
             .fetch_all(connection_mut)
             .await
-            .unwrap();
+            .int_err()?;
+        let last_event_id = rows.last().unwrap().event_id;
 
-        if let Some(last_row) = rows.last() {
-            Ok(EventID::new(last_row.event_id))
-        } else {
-            Ok(EventID::new(
-                i64::try_from(self.len().await.unwrap()).unwrap(),
-            ))
-        }
+        Ok(EventID::new(last_event_id))
     }
 
     async fn len(&self) -> Result<usize, InternalError> {
@@ -183,14 +182,13 @@ impl TaskSystemEventStore for TaskSystemEventStoreSqlite {
         pagination: TaskPaginationOpts,
     ) -> TaskIDStream {
         let mut tr = self.transaction.lock().await;
-        let dataset_id = dataset_id.clone();
+        let dataset_id = dataset_id.to_string();
 
         Box::pin(async_stream::stream! {
             let connection_mut = tr.connection_mut().await?;
 
-            let dataset_id_str = dataset_id.to_string();
-            let limit = i64::try_from(pagination.limit).unwrap();
-            let offset = i64::try_from(pagination.offset).unwrap();
+            let limit = i64::try_from(pagination.limit).int_err()?;
+            let offset = i64::try_from(pagination.offset).int_err()?;
 
             let mut query_stream = sqlx::query!(
                 r#"
@@ -199,7 +197,7 @@ impl TaskSystemEventStore for TaskSystemEventStoreSqlite {
                     WHERE dataset_id = $1 AND event_type = 'TaskEventCreated'
                     ORDER  BY task_id DESC LIMIT $2 OFFSET $3
                 "#,
-                dataset_id_str,
+                dataset_id,
                 limit,
                 offset,
             )