+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Flow system: implemented persistent repositories (PostgreSQL, SQLite) for flow configuration events
+
 ## [0.177.0] - 2024-04-25
 ### Added
 - REST data APIs and CLI commands now support different variations of JSON representation, including: